docs: Improve navigation and subprojects scope in global swh documentation
This commit adds a postprocessing step on the generated html documentation for swh.
Its purpose is to modify the indices and search scope in each swh subproject doc.

Prior to that, when browsing a subproject doc, indices and search were scoped
to all swh projects. This was not really convenient for the reader who only
wanted to focus on a specific subproject. Now indices and search are only scoped
to each swh subproject content.

Nevertheless, global indices and search are still available from the
root documentation index.