dependency graph: do not show extras as separate nodes
Currently, the dependency graph shows different nodes for, e.g., "swh.core" and
"swh.core[db]". With this change the two nodes are merged together.

While it is interesting to know which precise part of a module others depend
on, that is not the purpose of the dependency graph, which is meant to give
only a high-level view of dependencies.