Use latest sphinx version to build the doc
Now that sphinxcontrib-images extension has gained support for Sphinx > 2
in its 0.9.1 release, we can remove the restriction on Sphinx version in
the requirements file.

Also considering that we now rely on a virtualenv to build the doc, remove
the hacks to invocate sphinx-build and sphinx-apidoc in Makefiles.

Related to T2264