Make sub-toctrees visible
The sub-toctree pages are accessibly via the main toctree,
and if we make their content hidden, they are shown as empty pages.

Instead, this makes their content visible, so they list
their children.

This commit also removes the caption.
It wasn't shown so far because the toctrees were hidden,
but now it would be, but it actually looks a little silly to have
"Contents:" in bold just after a title.