tox/sphinx-dev: Workaround really long dependencies installation
The new pip resolver ends up backtracking numerous packages when
installing all depencies in a single pip command but this is
terribly slow and long to execute.

A workaround is to install each swh package one at a time by
executing pip multiple times, no backtracking happens in that
case and it reduces considerably the overall installation time.

Another advantage of proceeding like this is that it ensures each
swh package will be installed in development mode.

Also increase tox verbosity when calling make to have some feeback
of what it is doing.