sphinx/conf.py: Add standalone package doc configuration for tox build
Enable to build the local documentation of each swh package using a
dedicated tox environment.

In order to check for possibly introduced sphinx warnings in the
documentation sources, the documentation is built with the -W option
of sphinx-build.

As some references located in scopes external to the processed
package would not be resolved, related sphinx warnings are
suppressed to allow the doc build to succeed anyway.

Besides the errors related to references that can only be checked
when building the full documentation, all other sphinx errors
should be caught.

Related to T3258