Rename faq/faq.rst to faq/index.rst and deduplicate TOCs
The FAQ was put in a single file in faq/faq.rst as a hack, using
multiple top-level headers in the same file to make them appear
in the global TOC.

With this change, the TOC is moved to the index.rst file, with
a single top-level header, and uses a local TOC instead, as it
should be.

Additionally, this removed the manually-written TOC in the main
index to use an auto-generated one.