docs, user: Fix faq menu rendering glitch
As the toctree for the faq was included multiple times in index file,
the associated menu entry in the sidebar was not properly rendered.

In order to workaround the issue, ensure faq toctree is included once
in the index file and add reference to faq page in Getting started
top level section.

Also as faq subsection entries are no longer included in the sidebar
menu, display a local toctree at the top of the faq page.

Closes T3657