docs/Makefile: Fix local build of global swh documentation
It can be convenient to run the global swh documentation in the
current venv instead of using tox (notably to save time by skipping
dependencies installation).

It looks like the Makefile in docs folder was not really maintained
regarding that kind of documentation build so fix it in order to
restore local build.