Make “Run your own” tutorial more coherent
The tutorial on how to “run your own Software Heritage” felt quite dry.
Meanwhile, more details on how to start using the local
instance were already available in the documentation on how to use
the Docker environment.

So we pull in the relevant sections from
`swh-environment/docker/README.rst` to complement the existing tutorial.