Remove second-level headings from the dev doc + hide TOC
The only two titles at the second level were 'Important documentation links'
and 'Content', which look weird and are not very useful.

And this level prevented headings inside package documentation from being
shown in the TOC at all, which means the TOC couldn't be used to navigate
within a package's documentation.