Document takedown notices on mirrors
This document Software Heritage policy regarding how takedown notices
must be handled on mirrors. Currently we are missing tooling to
implement it in full, but we can still document the stopgap process
and what we are aiming.

Thanks to ardumont and vlorentz for proofreading and spotting several
typos, syntax and grammar issues.

Ref swh/meta#4975