distribute the glossary.rst file as package data
Distribute this docs/devel/glossary.rst file as package data in
share/swh-docs since it is needed for locally building the doc
from swh packages.

Adapt the add_glossary_to_index sphinx hook to look for this file in
share/swh-docs in addition to the current relative path (which only
works if swh-docs in installed in editable with proper pip
config-settings and who knows what else). For local/editable installed
swh-docs, look a bit more than previously to attempt to catch the case
of editable_mode=strict pip config